+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+* Bump: Sentry-Android to 5.7.0 and Sentry-Cocoa to 7.11.0 (#796)
+* Fix: Dio event processor safelly bails if no DioError in the exception list (#795)
+
 ## 6.5.0-alpha.1
 
 * Feat: Mobile Vitals - Native App Start (#749)